Fix unit tests
Most of the test build failures were due to changing the signature verifier from an interface to a function generated by a HOF.  This meant that the mocks in the tests had to be rewritten.
There are still some test failures in the smartbft package that I suspect haven’t worked in this version of Fabric.  They will be addressed in another commit.

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>